Reorganize code into files
