+++ src/main/java/Task.java
@@ -1,0 +1,27 @@
+public class Task {
+    private String description;
+    private boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getDescription() {
+        return this.description;
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void unmarkAsDone() {
+        this.isDone = false;
+    }
+
+    @Override
+    public String toString() {
+        String status = this.isDone ? "X" : " ";
+        return String.format("[%s] %s", status, this.description);
+    }
+}
